+++ src/xplot/init.c
@@ -5,9 +5,9 @@
 XContext	datacontext;
 
 MapMenu(val)
-int32_t val;
+int val;
 {
-int32_t menu_width;
+int menu_width;
 
     menu_width = 10*F->fontwidth;
     F->menu->mapped = val;
@@ -60,7 +60,7 @@ ChangeGeometry(geometry,frame)
 char    *geometry;
 Frame	*frame;
 {
-int32_t	status,x,y,width,height;
+int	status,x,y,width,height;
 XSizeHints hint;
 
     hint.flags = 0;
@@ -361,26 +361,26 @@ XSizeHints		hints;
 CreateWindows(frame)
 Frame		*frame;
 {
-int32_t	width;
-int32_t	height;
-int32_t	x;
-int32_t	y;
-uint32_t 		border;
-uint32_t 		background;
+int	width;
+int	height;
+int	x;
+int	y;
+unsigned long 		border;
+unsigned long 		background;
 XWindowAttributes	winfo;
 XSetWindowAttributes	attrib;
 Window			root_window;
-int32_t			borderwidth;
+int			borderwidth;
 XSizeHints		hints;
 WindowData		*windowdata;
 Graph			*graph;
 TextWindow		*text;
 MenuWindow		*menu;
-int32_t			status = 0;
-extern			int32_t frame_event();
-extern			int32_t graph_event();
-extern			int32_t text_event();
-extern			int32_t menu_event();
+int			status = 0;
+extern			int frame_event();
+extern			int graph_event();
+extern			int text_event();
+extern			int menu_event();
 char			*windowtitle;
 
     graph = frame->graph;
